# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.AirportResponse:
    properties:
      city:
        type: string
      code:
        type: string
      country:
        type: string
      id:
        type: string
      name:
        type: string
      timezone:
        type: string
    type: object
...
  dto.CreateFlightScheduleRequest:
    properties:
      arrival_airport_code:
        type: string
      arrival_time:
        type: string
      departure_airport_code:
        type: string
      departure_time:
        type: string
      flight_number:
        maxLength: 10
        type: string
      operating_days:
        type: string
    required:
    - arrival_airport_code
    - arrival_time
    - departure_airport_code
    - departure_time
    - flight_number
    - operating_days
    type: object
...
  dto.FlightScheduleResponse:
    properties:
      arrival_airport:
        $ref: '#/definitions/dto.AirportResponse'
      arrival_time:
        type: string
      departure_airport:
        $ref: '#/definitions/dto.AirportResponse'
      departure_time:
        type: string
      flight_number:
        type: string
      id:
        type: string
      operating_days:
        type: string
    type: object
...
  dto.InitiatePaymentRequest:
    properties:
      method:
        allOf:
        - $ref: '#/definitions/models.PaymentMethod'
        enum:
        - credit_card
        - debit_card
        - bank_transfer
        - e_wallet
      pnr_id:
        type: string
    required:
    - method
    - pnr_id
    type: object
...
  dto.PaymentResponse:
    properties:
      amount:
        type: number
      id:
        type: string
      method:
        $ref: '#/definitions/models.PaymentMethod'
      paid_at:
        type: string
      pnr_id:
        type: string
      status:
        $ref: '#/definitions/models.PaymentStatus'
    type: object
...
  dto.UpdateFlightScheduleRequest:
    properties:
      arrival_time:
        type: string
      departure_time:
        type: string
      operating_days:
        type: string
    type: object
...
  models.PaymentMethod:
    enum:
    - credit_card
    - debit_card
    - bank_transfer
    - e_wallet
    type: string
    x-enum-varnames:
    - PaymentMethodCreditCard
    - PaymentMethodDebitCard
    - PaymentMethodBankTransfer
    - PaymentMethodEWallet
  models.PaymentStatus:
    enum:
    - pending
    - success
    - failed
    - refunded
    type: string
    x-enum-varnames:
    - PaymentStatusPending
    - PaymentStatusSuccess
    - PaymentStatusFailed
    - PaymentStatusRefunded
...
  utils.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@airline.com
    name: API Support
  description: REST API untuk sistem Passenger Service System (PSS) maskapai penerbangan.
    Mencakup manajemen penerbangan, pemesanan tiket, check-in, bagasi, dan boarding
    pass.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Passenger Service Backend API
  version: "1.0"
paths:
...
  /payments:
    get:
      description: Mengambil semua riwayat pembayaran untuk sebuah PNR.
      parameters:
      - description: PNR UUID
        in: query
        name: pnr_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.PaymentResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Daftar pembayaran per PNR
      tags:
      - Payment
    post:
      consumes:
      - application/json
      description: 'Memulai proses pembayaran untuk sebuah PNR. Total dihitung dari
        harga kursi yang di-assign. Status awal: pending.'
      parameters:
      - description: Data pembayaran
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.InitiatePaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.PaymentResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
        "422":
          description: PNR sudah dibatalkan atau sudah ticketed
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Inisiasi pembayaran
      tags:
      - Payment
  /payments/{id}/confirm:
    post:
      consumes:
      - application/json
      description: Endpoint untuk payment gateway callback. Jika success=true, kursi
        diubah ke booked dan tiket diterbitkan. Hanya admin.
      parameters:
      - description: Payment UUID
        in: path
        name: id
        required: true
        type: string
      - description: Hasil pembayaran
        in: body
        name: request
        required: true
        schema:
          properties:
            success:
              type: boolean
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.PaymentResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
        "422":
          description: Pembayaran tidak dalam status pending
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Konfirmasi pembayaran (webhook)
      tags:
      - Payment
  /payments/{id}/refund:
    post:
      description: Memproses pengembalian uang untuk pembayaran yang sudah sukses.
        Hanya admin.
      parameters:
      - description: Payment UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.PaymentResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
        "422":
          description: Pembayaran belum success
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Proses refund
      tags:
      - Payment
...
  /schedules:
    get:
      description: Mengambil seluruh jadwal penerbangan. Bisa filter rute dengan ?dep=CGK&arr=DPS.
        Endpoint publik.
      parameters:
      - description: Kode IATA keberangkatan
        in: query
        name: dep
        type: string
      - description: Kode IATA tujuan
        in: query
        name: arr
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.FlightScheduleResponse'
                  type: array
              type: object
      summary: Daftar jadwal penerbangan
      tags:
      - FlightSchedule
    post:
      consumes:
      - application/json
      description: Membuat jadwal penerbangan baru (recurring schedule). Hanya admin.
      parameters:
      - description: Data jadwal
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateFlightScheduleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.FlightScheduleResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
        "409":
          description: Nomor penerbangan sudah ada
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Buat jadwal penerbangan
      tags:
      - FlightSchedule
  /schedules/{id}:
    delete:
      description: Menghapus jadwal penerbangan. Hanya admin.
      parameters:
      - description: Schedule UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Hapus jadwal penerbangan
      tags:
      - FlightSchedule
    get:
      description: Mengambil detail satu jadwal penerbangan. Endpoint publik.
      parameters:
      - description: Schedule UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.FlightScheduleResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Detail jadwal penerbangan
      tags:
      - FlightSchedule
    put:
      consumes:
      - application/json
      description: Memperbarui jadwal penerbangan (waktu, operating days). Hanya admin.
      parameters:
      - description: Schedule UUID
        in: path
        name: id
        required: true
        type: string
      - description: Data update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateFlightScheduleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.FlightScheduleResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Update jadwal penerbangan
      tags:
      - FlightSchedule
...
securityDefinitions:
  BearerAuth:
    description: 'Masukkan token dengan format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"